New Core fixes
git-svn-id: svn://ultimatepp.org/upp/trunk@9771 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/plugin/lzma/lzma.cpp
@@ -31,7 +31,7 @@ struct IUppInStream {
 			*size = 0;
 			return SZ_ERROR_READ;
 		}
-		*size = h;
+		*size = (size_t)h;
 		return SZ_OK;
 	}
